absl.FLAGS added and dependence from  DifferentiableHOS removed
