+++ scripts/fit_pgd.py
@@ -1,96 +1,33 @@
 import numpy as np
 import matplotlib.pyplot as plt
-
 import tensorflow as tf
-import numpy as np
 import sys
 sys.path.append('../../flowpm/')
-sys.path.append('../../../DifferentiableHOS/')
-sys.path.append('../../DifferentiableHOS/')
-# import DifferentiableHOS as DHOS
-import flowpm
 import pickle
+import flowpm
 import flowpm.tfpower as tfpower
 import flowpm.scipy.interpolate as interpolate
 from flowpm.tfpower import linear_matter_power
-# import jax
+from absl import app
+from absl import flags
 from flowpm import tfpm
 import jax_cosmo as jc
-from DifferentiableHOS.pk import pk as pkl
-from DifferentiableHOS.pk import _initialize_pk as initpk
+from flowpm.pk import pk as pkl
+from flowpm.pk import _initialize_pk as initpk
 import jax_cosmo.power as power
-
-cosmology = flowpm.cosmology.Planck15()
-
 from flowpm.utils import white_noise, c2r3d, r2c3d, cic_paint, cic_readout
-fftk = flowpm.kernels.fftk
-laplace_kernel = flowpm.kernels.laplace_kernel
-gradient_kernel = flowpm.kernels.gradient_kernel
+
+flags.DEFINE_string("filename", "results_fit_PGD.pkl", "Output filename")
+flags.DEFINE_float("box_size", 64.,
+                   "Transverse comoving size of the simulation volume")
+flags.DEFINE_integer("nc", 64,
+                     "Number of transverse voxels in the simulation volume")  
+flags.DEFINE_integer("nsteps", 40, "Number of steps in the N-body simulation")
+flags.DEFINE_integer("B", 1, "Scale resolution factor")
 
 
-box_size = 200.
-nc = 128       
-nsteps= 10
-B=1
-pm_nc_factor = B
-ia = -1 #Scale factor to fit for
-niters = 50
+FLAGS = flags.FLAGS
 
-# Create some initial conditions
-klin = tf.constant(np.logspace(-4, 1, 512), dtype=tf.float32)
-pk = linear_matter_power(cosmology, klin)
-pk_fun = lambda x: tf.cast(tf.reshape(interpolate.interp_tf(tf.reshape(tf.cast(x, tf.float32), [-1]), klin, pk), x.shape), tf.complex64)
-initial_conditions = flowpm.linear_field(
-      [nc, nc, nc],
-      [box_size, box_size, box_size],
-      pk_fun,
-      batch_size=1)
-initial_state = flowpm.lpt_init(cosmology, initial_conditions, 0.1)
-
-stages = np.linspace(0.1, 1., nsteps, endpoint=True)
-# Run the Nbody
-states2 = flowpm.nbody(cosmology,
-                        initial_state,
-                    stages, [nc, nc, nc],
-                    pm_nc_factor= B,
-                    return_intermediate_states=True)
-
-print('Simulation done')
-
-#Nyquist and number of modes in case we need weight for PGD fitting loss
-kny = np.pi*nc/box_size
-Nmodes = initpk(shape=initial_state.shape[1:],boxsize=np.array([box_size, box_size,
-                                            box_size]),kmin=np.pi/box_size,dk=2*np.pi/box_size)[1].numpy().real[1:-1]
-
-
-#######################################
-#######################################
-##Estimate PS for the last time snapshot
-pk_array=[]
-pk_array1=[]
-
-# for i in range(len(states2)):
-for i in [ia]:
-    print(i)
-    final_field1 = flowpm.cic_paint(tf.zeros_like(initial_conditions), states2[i][1][0])
-    final_field1 = tf.reshape(final_field1, [nc, nc, nc])
-    k, power_spectrum1 = pkl(final_field1,shape=final_field1.shape,boxsize=np.array([box_size, box_size,
-                                         box_size]),kmin=np.pi/box_size,dk=2*np.pi/box_size)
-    pk_array1.append(power_spectrum1)
-
-    
-pk_jax=[]
-cosmo=jc.Planck15()
-# for i in range(len(states2)):
-for i in [ia]:
-    print(i)
-    pk_jax.append(power.nonlinear_matter_power(cosmo, k, states2[i][0]))
-
-print(len(k))
-print(len(pk_jax[-1]))
-print('PS computed')
-
-#######################################
 #######################################
 #Same correction function as flowpm but writing as tf.function
 @tf.function
@@ -100,20 +37,18 @@ def PGD_correction(
     """                                                                                                                                                                                                                                                                                                                          """
     print('new graph')
     state = tf.convert_to_tensor(state, name="state")
-
+    pm_nc_factor=FLAGS.B
     shape = state.get_shape()
     batch_size = shape[1]
-    ncpm = [nc, nc, nc]
-    ncf = [nc * pm_nc_factor]*3
-#     ncf = [n * pm_nc_factor for n in nc]
-
+    ncpm = [FLAGS.nc, FLAGS.nc, FLAGS.nc]
+    ncf = [FLAGS.nc * pm_nc_factor]*3
     rho = tf.zeros([batch_size] + ncf)
     wts = tf.ones((batch_size, ncpm[0] * ncpm[1] * ncpm[2]))
     nbar = ncpm[0] * ncpm[1] * ncpm[2] / (ncf[0] * ncf[1] * ncf[2])
 
     rho = cic_paint(rho, tf.multiply(state[0], pm_nc_factor), wts)
     rho = tf.multiply(rho,
-                      1. / nbar)  # I am not sure why this is not needed here                                                                                                                                                                                                                                                 
+                      1. / nbar)                  
     delta_k = r2c3d(rho, norm=ncf[0] * ncf[1] * ncf[2])
     alpha, kl, ks = tf.split(pgdparams, 3)
     update = tfpm.apply_PGD(tf.multiply(state[0], pm_nc_factor),
@@ -127,79 +62,136 @@ def PGD_correction(
 
 
 @tf.function
-def pgd(params):
-#     alpha, kl, ks = tf.split(params, 3)
+def pgd(params,states,initial_conditions):
     print('pgd graph')
-    dx = PGD_correction(states2[ia][1], params)
-    new_state= states2[ia][1][0]+dx
+    ia = -1 
+    dx = PGD_correction(states[ia][1], params)
+    new_state= states[ia][1][0]+dx
     final_field = flowpm.cic_paint(tf.zeros_like(initial_conditions), new_state)
-    final_field=tf.reshape(final_field, [nc, nc, nc])
-    k, power_spectrum = pkl(final_field,shape=final_field.shape,boxsize=np.array([box_size, box_size,
-                                            box_size]),kmin=np.pi/box_size,dk=2*np.pi/box_size)
+    final_field=tf.reshape(final_field, [FLAGS.nc, FLAGS.nc, FLAGS.nc])
+    k, power_spectrum = pkl(final_field,shape=final_field.shape,boxsize=np.array([FLAGS.box_size, FLAGS.box_size,
+                                            FLAGS.box_size]),kmin=np.pi/FLAGS.box_size,dk=2*np.pi/FLAGS.box_size)
 
     return k, power_spectrum, final_field
 
-
 #######################################
-#######################################
-#Finally fit
-#Loss 
-#weight = Nmodes**0.5  * (1-k/kny)
-weight = 1-k/kny
-weight = weight/weight.sum()
-weight = tf.cast(weight, tf.float32)
-# weight = 1.
-
 @tf.function
-def get_grads(params):
+def get_grads(params,states,initial_conditions,pk_jax,weight):
+    ia = -1 
     with tf.GradientTape() as tape:
         tape.watch(params)
-        k, ps, field = pgd(params)
-#         loss = tf.reduce_sum(ps)
-        psref = tf.constant(np.array(pk_jax[ia]))
+        k, ps, field = pgd(params,states,initial_conditions)
+        psref=tf.convert_to_tensor(pk_jax[ia])
         loss = tf.reduce_sum((weight*(1 - ps/psref))**2)
     grad = tape.gradient(loss, params)
     return loss, grad
 
+def main(_):
+    
+    cosmology = flowpm.cosmology.Planck15()
+    fftk = flowpm.kernels.fftk
+    laplace_kernel = flowpm.kernels.laplace_kernel
+    gradient_kernel = flowpm.kernels.gradient_kernel
+    ia = -1 #Scale factor to fit for
+    niters = 50
+    # Create some initial conditions
+    klin = tf.constant(np.logspace(-4, 1, 512), dtype=tf.float32)
+    pk = linear_matter_power(cosmology, klin)
+    pk_fun = lambda x: tf.cast(tf.reshape(interpolate.interp_tf(tf.reshape(tf.cast(x, tf.float32), [-1]), klin, pk), x.shape), tf.complex64)
+    initial_conditions = flowpm.linear_field(
+          [FLAGS.nc, FLAGS.nc, FLAGS.nc],
+          [FLAGS.box_size, FLAGS.box_size, FLAGS.box_size],
+          pk_fun,
+          batch_size=1)
+    initial_state = flowpm.lpt_init(cosmology, initial_conditions, 0.1)
+    stages = np.linspace(0.1, 1., FLAGS.nsteps, endpoint=True)
+    # Run the Nbody
+    states2 = flowpm.nbody(cosmology,
+                            initial_state,
+                        stages, [FLAGS.nc, FLAGS.nc, FLAGS.nc],
+                        pm_nc_factor= FLAGS.B,
+                        return_intermediate_states=True)
+    print('Simulation done')
+    #Nyquist and number of modes in case we need weight for PGD fitting loss
+    kny = np.pi*FLAGS.nc/FLAGS.box_size
+    Nmodes = initpk(shape=initial_state.shape[1:],boxsize=np.array([FLAGS.box_size, FLAGS.box_size,
+                                        FLAGS.box_size]),kmin=np.pi/FLAGS.box_size,dk=2*np.pi/FLAGS.box_size)[1].numpy().real[1:-1]
+    
+    ##Estimate PS for the last time snapshot
+    pk_array=[]
+    for i in [ia]:
+        print(i)
+        final_field1 = flowpm.cic_paint(tf.zeros_like(initial_conditions), states2[i][1][0])
+        final_field1 = tf.reshape(final_field1, [FLAGS.nc, FLAGS.nc, FLAGS.nc])
+        k, power_spectrum1 = pkl(final_field1,shape=final_field1.shape,boxsize=np.array([FLAGS.box_size, FLAGS.box_size,
+                                             FLAGS.box_size]),kmin=np.pi/FLAGS.box_size,dk=2*np.pi/FLAGS.box_size)
+        pk_array.append(power_spectrum1)
+    pk_jax=[]
+    cosmo=jc.Planck15()
 
-kl, ks = 0.3, 12
-alpha = 0.3 
-params0 = np.array([alpha, kl, ks]).astype(np.float32)
-params = tf.Variable(name='pgd', shape=(3), dtype=tf.float32,
-                             initial_value=params0, trainable=True)
-print('Init params : ', params)
+    for i in [ia]:
+        print(i)
+        pk_jax.append(power.nonlinear_matter_power(cosmo, k, states2[i][0]))
 
-opt = tf.keras.optimizers.Adam(learning_rate=0.1)
-losses, pparams = [], []
-print(params0)
-for i in range(niters):
-    loss, grads = get_grads(params)
-    opt.apply_gradients(zip([grads], [params]))
-    print(i, loss, params)
-    losses.append(loss)
-    pparams.append(params.numpy().copy())
+    print(len(k))
+    print(len(pk_jax[-1]))
+    print('PS computed') 
+    #Finally fit
+    #Loss 
+    weight = 1-k/kny
+    weight = weight/weight.sum()
+    weight = tf.cast(weight, tf.float32)
+    kl, ks = 0.3, 12
+    alpha = 0.3 
+    params0 = np.array([alpha, kl, ks]).astype(np.float32)
+    params = tf.Variable(name='pgd', shape=(3), dtype=tf.float32,
+                                 initial_value=params0, trainable=True)
+    print('Init params : ', params)
 
-print('Final params : ', params)
-#Plot fit params
-plt.plot(losses)
-plt.semilogy()
-plt.grid(which='both')
-plt.savefig('losses.png')
-plt.close()
+    opt = tf.keras.optimizers.Adam(learning_rate=0.1)
+    losses, pparams = [], []
+    print(params0)
+    for i in range(niters):
+        loss, grads = get_grads(params,states2,initial_conditions,pk_jax,weight)
+        opt.apply_gradients(zip([grads], [params]))
+        print(i, loss, params)
+        losses.append(loss)
+        pparams.append(params.numpy().copy())
+
+    print('Final params : ', params)
+    #Plot fit params
+    plt.plot(losses)
+    plt.semilogy()
+    plt.grid(which='both')
+    plt.savefig('losses.png')
+    plt.close()
 
 
-k, ps, field = pgd(params)
-k, ps0, field0 = pgd(tf.constant(params0))
-plt.plot(k, pk_array1[ia]/pk_jax[ia], "k:", label='B=1')
-plt.plot(k, ps0/pk_jax[ia], 'k--', label='0')
-plt.semilogx()
-plt.grid(which='both')
-# plt.fill_between(k, k*0 + 1 + 1/Nmodes**0.5, k*0 + 1 - 1/Nmodes**0.5)
-for i in range(niters/10.):
-    j = 10*i + 9
-    try:
-        k, ps, field = pgd(tf.constant(pparams[j]))
-        plt.plot(k, ps/pk_jax[ia], '-', label=j)
-    except: pass
-plt.legend()
-plt.savefig('pgdfit.png')
+    k, ps, field = pgd(params,states2,initial_conditions)
+    k, ps0, field0 = pgd(tf.constant(params0),states2,initial_conditions)
+    plt.plot(k, pk_array[ia]/pk_jax[ia], "k:", label='B=1')
+    plt.plot(k, ps0/pk_jax[ia], 'k--', label='0')
+    plt.semilogx()
+    plt.grid(which='both')
+    for i in range(niters//10):
+        j = 10*i + 9
+        try:
+            k, ps, field = pgd(tf.constant(pparams[j]),states2,initial_conditions)
+            plt.plot(k, ps/pk_jax[ia], '-', label=j)
+        except: pass
+    plt.legend()
+    plt.savefig('pgdfit.png')
+    pickle.dump(
+      {
+          'B': FLAGS.B,
+          'nsteps': FLAGS.nsteps,
+          'pgdparams':params.numpy(),
+          'alpha': params.numpy()[0],
+          'kl': params.numpy()[1],
+          'ks': params.numpy()[2],
+      }, open(FLAGS.filename, "wb"))
+    
+    
+if __name__ == "__main__":
+  app.run(main)
+
